% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plots.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split Violin Plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes()}

\item{data}{A dataframe}

\item{stat}{The statistical transformation to use on the data for this layer,}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{...}{other arguments passed on to layer. These are often aesthetics,
used to set an aesthetic to a fixed value, like colour = "red" or size = 3.
They may also be parameters to the paired geom/stat.}

\item{draw_quantiles}{numeric vector of length two, specifying which
quantiles to draw as segments. The default is to draw the quartiles
(0.25, 0.75). To suppress drawing quantiles, set this to numeric(0).}

\item{trim}{If TRUE (default), trim the tails of the violins to the range
of the data. If FALSE, don't trim the tails.}

\item{scale}{if "area" (default), all violins have the same area
(before trimming the tails). If "count", areas are scaled proportionally to
the number of observations. If "width", all violins have the same maximum
width.}

\item{na.rm}{logical (default is FALSE) remove NA values from the data before
computing statistics.}

\item{show.legend}{logical. Should this layer be included in the legends?
NA, the default, includes if any aesthetics are mapped. FALSE never includes,
and TRUE always includes. It can also be a named logical vector to finely
select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them. This is most useful for helper functions that define
both data and aesthetics and shouldn't inherit behaviour from the default
plot specification, e.g. borders().}
}
\value{
A ggplot2 geom
}
\description{
A split violin plot allows to display the KDE of two groups
next to each other. It is a variation of the violin plot. Code was copied
from this href{https://stackoverflow.com/questions/35717353/split-violin-plot-with-ggplot2}{Stackoverflow post}.
The parameter descriptions coincide with the official ones of the geom_violin
layer
}
