% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{label_facets}
\alias{label_facets}
\title{Label facets of a ggplot2 plot
Function adapted from the tag_facet function in the egg package}
\usage{
label_facets(
  plot,
  type = c("alpha", "ALPHA", "num", "roman", "ROMAN"),
  open = "(",
  close = ")",
  vjust = 2,
  hjust = -0.5,
  top_down = FALSE
)
}
\arguments{
\item{plot}{A ggplot2 plot}

\item{type}{type of label to use: select one from "alpha", "ALPHA", "num",
"roman", "ROMAN"}

\item{open}{string/character preceeding type. Default is "("}

\item{close}{string/character following type. Default is ")"}

\item{vjust}{vertical alignment of label. Default is 2}

\item{hjust}{horizontal alignment of label. Default is -0.5}

\item{top_down}{logical. Default is FALSE. If TRUE, labels are ordered from
top to bottom, if FALSE, labels are ordered from left to right}
}
\value{
A ggplot2 plot with labelled facets
}
\description{
Label facets of a ggplot2 plot
Function adapted from the tag_facet function in the egg package
}
