% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{scale_color_paper}
\alias{scale_color_paper}
\title{Custom Color Scale for Papers}
\usage{
scale_color_paper(...)

scale_color_paper(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{discrete_scale}}}
}
\value{
None

None
}
\description{
Custom Color Scale for Papers

Custom Fill Scale for Papers
}
