% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{scale_fill_paper}
\alias{scale_fill_paper}
\title{Custom Fill Scale for Papers}
\usage{
scale_fill_paper(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{discrete_scale}}}
}
\value{
None
}
\description{
Custom Fill Scale for Papers
}
