% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{set_paper_plot_specs}
\alias{set_paper_plot_specs}
\title{Set plot specifications}
\usage{
set_paper_plot_specs(
  fontfamily = "ibm",
  plot_background_fill = NA,
  panel_background_fill = NA,
  nseg = 12,
  dpi = 300
)
}
\arguments{
\item{fontfamily}{font family to use for the plots}

\item{plot_background_fill}{background fill of the plot}

\item{panel_background_fill}{background fill of the panel}

\item{nseg}{control the smoothness of the glyphs}

\item{dpi}{dpi of the text}
}
\description{
This function sets the plot specifications for the current
session. It sets the theme for plots, enables showtext for the graphic
device and sets the dpi of the graphic device.
}
