% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{paper_theme}
\alias{paper_theme}
\title{Paper Plot Theme}
\usage{
paper_theme(
  fontfamily = "IBMPlexSans-Medium",
  plot_background_fill = NA,
  panel_background_fill = NA
)
}
\arguments{
\item{fontfamily}{Fontfamily that will be used. Default is IBMPlexSans-Medium.
Next to the traditional fonts the LaTeX/ Latin Roman 10 font will be loaded
and
can be selected using "lmroman" as fontfamily}

\item{plot_background_fill}{fill color of the plot background; default is NA}

\item{panel_background_fill}{fill color of the panel background; default is
NA}
}
\value{
None
}
\description{
Set custom theme for plots that is well suited for papers
}
