% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palettes.R
\name{scale_color_paper}
\alias{scale_color_paper}
\title{Custom Color Scale for Papers}
\usage{
scale_color_paper(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{discrete_scale}}}
}
\value{
None
}
\description{
A simple pastel color palette for use in papers. Other packages that
include useful color palettes are \code{RColorBrewer}, \code{viridis},
\code{wesanderson}, and \code{ggsci}.
}
