% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palettes.R
\name{scale_fill_paper_bw}
\alias{scale_fill_paper_bw}
\title{Black and White Fill Scale for Papers}
\usage{
scale_fill_paper_bw(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{discrete_scale}}}
}
\value{
None
}
\description{
A black and white color palette for use in papers.
}
