% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{set_plot_specs}
\alias{set_plot_specs}
\title{Set plot specifications}
\usage{
set_plot_specs(dpi = 600)
}
\arguments{
\item{dpi}{dpi of the text}
}
\description{
This function sets the plot specifications for the current
session. It sets the theme for plots, enables showtext for the graphic
device and sets the dpi of the graphic device.
}
